+++ setup.py
@@ -24,6 +24,7 @@ setup(
         "microcosm-caching>=0.2.0",
         "microcosm-daemon>=1.0.0",
         "microcosm-logging>=1.3.0",
+        "tornado<6",
     ],
     extras_require={
         "metrics": "microcosm-metrics>=2.5.0",
